requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
distro==1.9.0
faiss-cpu==1.13.2
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
numpy==2.4.4
openai==2.32.0
packaging==26.1
pydantic==2.13.3
pydantic_core==2.46.3
pypdf==6.10.2
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
sniffio==1.3.1
starlette==1.0.0
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
pypdf>=4.0.0